Change selection mode in data grid
The selected column has been removed. Instead we rely on the DataGrid
multi-selection capabilities. This allows to select multiple items
using keyboard keys Shift and Control, and also avoids the multiple
clicks issue to select a CheckBox.

Showing details for a node is now triggered by double clicking on a
row. This allows to select multiple items without showing the details
for all of them.

Also, the style of the Apply button is improved to look more like
Dynamo.